+++ features/settings/src/main/kotlin/voice/features/settings/SettingsListener.kt
@@ -1,5 +1,6 @@
 package voice.features.settings
 
+import voice.core.data.LockscreenSecondaryTextMode
 import voice.core.data.LockscreenSliderMode
 import voice.core.data.MediaButtonClickAction
 import java.time.LocalTime
@@ -25,6 +26,8 @@ interface SettingsListener {
   fun openHiddenBooks()
   fun openBackup()
   fun openLicenses()
+  fun suggestIdea()
+  fun reportProblem()
 
   fun onMediaButtonDoubleClickRowClick()
   fun onMediaButtonTripleClickRowClick()
@@ -33,6 +36,8 @@ interface SettingsListener {
 
   fun onLockscreenSliderRowClick()
   fun setLockscreenSliderMode(mode: LockscreenSliderMode)
+  fun onLockscreenSecondaryTextRowClick()
+  fun setLockscreenSecondaryTextMode(mode: LockscreenSecondaryTextMode)
 
   fun setExperimentalPlaybackPersistence(enabled: Boolean)
   fun onExperimentalPlaybackPersistenceInfoClick()
@@ -65,6 +70,8 @@ interface SettingsListener {
       override fun openHiddenBooks() {}
       override fun openBackup() {}
       override fun openLicenses() {}
+      override fun suggestIdea() {}
+      override fun reportProblem() {}
 
       override fun onMediaButtonDoubleClickRowClick() {}
       override fun onMediaButtonTripleClickRowClick() {}
@@ -73,6 +80,8 @@ interface SettingsListener {
 
       override fun onLockscreenSliderRowClick() {}
       override fun setLockscreenSliderMode(mode: LockscreenSliderMode) {}
+      override fun onLockscreenSecondaryTextRowClick() {}
+      override fun setLockscreenSecondaryTextMode(mode: LockscreenSecondaryTextMode) {}
 
       override fun setExperimentalPlaybackPersistence(enabled: Boolean) {}
       override fun onExperimentalPlaybackPersistenceInfoClick() {}